Software/Pls-Dont-Burn/requirements.txt
sounddevice==0.5.3
numpy==2.3.4
python-dotenv==1.2.1
google-genai==1.49.0
librosa==0.11.0
soundfile==0.12.1
